backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.4
pyodbc==5.1.0
psycopg2-binary==2.9.9
pymysql==1.1.1
sqlalchemy==2.0.31
python-dotenv==1.0.1
